Support both query and queries server
