--- src/query.hh
@@ -1,45 +1,0 @@
-#pragma once
-
-#include <iostream>
-#include <fstream>
-#include <utility>
-#include <algorithm>
-#include <string>
-#include <vector>
-#include <map>
-#include <queue>
-#include <set>
-#include <unordered_set>
-#include <set>
-#include <bitset>
-#include <cassert>
-#include <fstream>
-
-#include <time.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <stdlib.h>
-#include <fcntl.h>
-#include <unistd.h>
-#include <sys/resource.h>
-#include <sys/stat.h>
-#include <sys/time.h>
-#include <sys/mman.h>
-#include <openssl/rand.h>
-
-#include "MantisFS.h"
-#include "ProgOpts.h"
-#include "spdlog/spdlog.h"
-#include "kmer.h"
-#include "coloreddbg.h"
-#include "common_types.h"
-#include "CLI/CLI.hpp"
-#include "CLI/Timer.hpp"
-
-namespace query 
-{
-  ColoredDbg<SampleObject<CQF<KeyObject>*>, KeyObject> * colored_dbg; 
-  spdlog::logger * console; 
-
-  void run_query(std::string query_file, std::string output_file); 
-}
